using 64 bits for value of puk
git-svn-id: https://vsmartcard.svn.sourceforge.net/svnroot/vsmartcard@632 96b47cad-a561-4643-ad3b-153ac7d7599c

+++ npa/src/npa-tool.c
@@ -343,8 +343,8 @@ main (int argc, char **argv)
     if (dobreak) {
         /* The biggest buffer sprintf could write with "%llu" */
         char secretbuf[strlen("18446744073709551615")+1];
-        unsigned long long secret = 0;
-        unsigned long long maxsecret = 0;
+        uint64_t secret = 0;
+        uint64_t maxsecret = 0;
 
         if (usepin) {
             pace_input.pin_id = PACE_PIN;
